requirements_plugin.txt
pytest-playwright==0.4.3
pytest-xdist==3.4.0